pyproject.toml
[project]
name = "music-production"
version = "0.1.0"
description = "Learning Ableton Live through Python with pylive"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pylive>=0.4.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9",
]

[tool.ruff]
line-length = 99
target-version = "py312"

[tool.ruff.lint]
select = ["D", "E", "F", "I", "W"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"config.py" = ["D"]  # constants-only file, no docstrings needed

[tool.ruff.format]
quote-style = "double"
